# openapi.yaml
...
openapi: 3.1.0
info:
  title: openagentix API
  version: 0.1.0
  description: "Control node API of openagentix: agents, events, runs, approvals, audit, costs and RBAC."
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Session or API token (`oax_...`)
    runToken:
      type: http
      scheme: bearer
      description: Signed run token (`oaxrt....`) of a worker node
    webhookSignature:
      type: apiKey
      in: header
      name: x-oax-signature
      description: HMAC-SHA256 signature
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: stable error code
        message:
          type: string
        details: {}
paths:
...
  /v1/auth/login:
    post:
      summary: Log in with local or LDAP credentials
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 1
                  maxLength: 320
                password:
                  type: string
                  minLength: 1
                  maxLength: 1000
                method:
                  default: auto
                  type: string
                  enum:
                    - auto
                    - local
                    - ldap
              required:
                - username
                - password
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expiresAt:
                    type: string
                    description: ISO 8601 timestamp
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      tenantId:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      email:
                        type: string
                      displayName:
                        type: string
                      source:
                        type: string
                      globalRoles:
                        type: array
                        items:
                          type: string
                      disabled:
                        type: boolean
                      teams:
                        type: array
                        items:
                          type: object
                          properties:
                            teamId:
                              type: string
                              format: uuid
                              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                            role:
                              type: string
                          required:
                            - teamId
                            - role
                          additionalProperties: false
                      createdAt:
                        type: string
                        description: ISO 8601 timestamp
                      lastLoginAt:
                        anyOf:
                          - type: string
                            description: ISO 8601 timestamp
                          - type: "null"
                    required:
                      - id
                      - tenantId
                      - email
                      - displayName
                      - source
                      - globalRoles
                      - disabled
                      - teams
                      - createdAt
                      - lastLoginAt
                    additionalProperties: false
                required:
                  - token
                  - expiresAt
                  - user
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  details: {}
                required:
                  - error
                  - message
                additionalProperties: false
...
  /v1/auth/logout:
    post:
      summary: Revoke the current session token
      tags:
        - auth
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/auth/oidc/callback:
    get:
      summary: OIDC redirect target; returns a session or redirects to the UI
      tags:
        - auth
      responses:
        "200":
          description: Default Response
        "302":
          description: Redirect to `<OAX_UI_URL>/auth/callback#token=…&expiresAt=…` (or 200 JSON without a UI URL)
          headers:
            Location:
              schema:
                type: string
  /v1/me:
    get:
      summary: Current principal
      tags:
        - auth
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      tenantId:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      email:
                        type: string
                      displayName:
                        type: string
                      source:
                        type: string
                      globalRoles:
                        type: array
                        items:
                          type: string
                      disabled:
                        type: boolean
                      teams:
                        type: array
                        items:
                          type: object
                          properties:
                            teamId:
                              type: string
                              format: uuid
                              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                            role:
                              type: string
                          required:
                            - teamId
                            - role
                          additionalProperties: false
                      createdAt:
                        type: string
                        description: ISO 8601 timestamp
                      lastLoginAt:
                        anyOf:
                          - type: string
                            description: ISO 8601 timestamp
                          - type: "null"
                    required:
                      - id
                      - tenantId
                      - email
                      - displayName
                      - source
                      - globalRoles
                      - disabled
                      - teams
                      - createdAt
                      - lastLoginAt
                    additionalProperties: false
                  tenant:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      slug:
                        type: string
                      name:
                        type: string
                    required:
                      - id
                      - slug
                      - name
                    additionalProperties: false
                  platformAdmin:
                    type: boolean
                  kind:
                    type: string
                    enum:
                      - user
                      - token
                  permissions:
                    type: array
                    items:
                      type: string
                  bindings:
                    type: array
                    items:
                      type: object
                      properties:
                        role:
                          type: string
                        teamId:
                          anyOf:
                            - type: string
                              format: uuid
                              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                            - type: "null"
                      required:
                        - role
                        - teamId
                      additionalProperties: false
                required:
                  - user
                  - tenant
                  - platformAdmin
                  - kind
                  - permissions
                  - bindings
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
servers:
  - url: http://localhost:8080
    description: local development